Enable exec inherits.
