+++ examples/Orga/config.yaml
@@ -107,7 +107,7 @@ paras:
         - 啊，可能是他们把火星的战力都调到那里去了。
         - 不过这也和我们没关系了。
         - 心情很愉快啊。
-        - /rd/加把劲骑士/那是当然的了！大家也有救了，\res{ch_tkk}又那么努力，我也\ruby{要加油了}{是加把劲骑士}！
+        - /rd/加把劲骑士/那是当然的了！大家也有救了，\{\}又那么努力，我也\ruby{要加油了}{是加把劲骑士}！
         - 啊。
         - 没错，我们至今为止所做的一切，并不是全部徒劳的。今后也是，只要我们不停下脚步，道路就会不断延伸。
     -
