+++ plugins/format/Cargo.toml
@@ -10,5 +10,6 @@ forced-target = "wasm32-unknown-unknown"
 crate-type = ["cdylib"]
 
 [dependencies]
+gal-primitive = { path = "../../utils/gal-primitive", features = ["rt-format"] }
 gal-bindings = { path = "../../utils/gal-bindings" }
 rt-format = "0.3"
