+++ utils/gal-primitive/src/lib.rs
@@ -216,6 +216,78 @@ impl Serialize for RawValue {
     }
 }
 
+#[cfg(feature = "rt-format")]
+use rt_format::{Format, FormatArgument, Specifier};
+
+#[cfg(feature = "rt-format")]
+impl FormatArgument for RawValue {
+    fn supports_format(&self, specifier: &Specifier) -> bool {
+        match self {
+            RawValue::Unit | RawValue::Bool(_) | RawValue::Str(_) => {
+                matches!(specifier.format, Format::Debug | Format::Display)
+            }
+            RawValue::Num(_) => true,
+        }
+    }
+
+    fn fmt_display(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        use std::fmt::Display;
+        match self {
+            RawValue::Unit => Ok(()),
+            RawValue::Bool(b) => b.fmt(f),
+            RawValue::Num(n) => n.fmt(f),
+            RawValue::Str(s) => s.fmt(f),
+        }
+    }
+
+    fn fmt_debug(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        use std::fmt::Debug;
+        self.fmt(f)
+    }
+
+    fn fmt_octal(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            RawValue::Num(n) => std::fmt::Octal::fmt(n, f),
+            _ => Err(std::fmt::Error),
+        }
+    }
+
+    fn fmt_lower_hex(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            RawValue::Num(n) => std::fmt::LowerHex::fmt(n, f),
+            _ => Err(std::fmt::Error),
+        }
+    }
+
+    fn fmt_upper_hex(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            RawValue::Num(n) => std::fmt::UpperHex::fmt(n, f),
+            _ => Err(std::fmt::Error),
+        }
+    }
+
+    fn fmt_binary(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            RawValue::Num(n) => std::fmt::Binary::fmt(n, f),
+            _ => Err(std::fmt::Error),
+        }
+    }
+
+    fn fmt_lower_exp(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            RawValue::Num(n) => std::fmt::LowerExp::fmt(n, f),
+            _ => Err(std::fmt::Error),
+        }
+    }
+
+    fn fmt_upper_exp(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            RawValue::Num(n) => std::fmt::UpperExp::fmt(n, f),
+            _ => Err(std::fmt::Error),
+        }
+    }
+}
+
 #[cfg(test)]
 mod test {
     use crate::*;
