+++ utils/gal-runtime/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+gal-primitive = { path = "../gal-primitive", features = ["rt-format"] }
 gal-bindings-types = { path = "../gal-bindings-types" }
 gal-script = { path = "../gal-script" }
 fallback = { path = "../fallback" }
@@ -16,6 +17,7 @@ anyhow = "1.0"
 wasmer = "2.3"
 wasmer-wasi = "2.3"
 unicode-width = "0.1"
+rt-format = "0.3"
 tokio = { version = "1", features = ["fs", "macros"] }
 tokio-stream = { version = "0.1", features = ["fs"] }
 stream-future = "0.3"
